+++ src/scripts/ternaryData.js
@@ -1,6 +1,7 @@
 import ternary from "./ternary";
 
 
+
 const ternaryData = {
 
     connectToData(fetchObject) {
